+++ Saturn/src/Saturn/ImGui/ContentBrowserPanel/ContentBrowserItem.cpp
@@ -252,7 +252,7 @@ namespace Saturn {
 
 					if( m_Asset )
 					{
-						ImGui::Text( "Asset: %" PRIu64, m_Asset->ID );
+						ImGui::Text( "Asset: %" PRIu64, ( uint64_t ) m_Asset->ID );
 						ImGui::Text( "Asset Name: %s", m_Asset->Name.c_str() );
 						ImGui::Text( "Asset Version: %" PRIu64, m_Asset->Version );
 					}
